ucd-msc/final-project/blackteam/web/requirements.txt
coverage==3.7.1
coveralls==0.5
Django==1.8.1
docopt==0.6.2
flake8==2.4.0
gunicorn==19.3.0
mccabe==0.3
pep8==1.5.7
psycopg2==2.6
pyflakes==0.8.1
PyYAML==3.11
redis==2.10.3
requests==2.7.0
selenium==2.46.0
wheel==0.24.0
feedparser==5.2.0.post1
tweepy==3.3.0
authomatic==0.0.13
celery==3.1.18
djangorestframework==3.1
nltk==3.0
numpy==1.9.2
